# Specify files for port detection for every detector