--- a/pkg/apis/enricher/framework/dotnet/dotnet_detector.go
+++ b/pkg/apis/enricher/framework/dotnet/dotnet_detector.go
@@ -31,6 +31,11 @@ func (d DotNetDetector) GetSupportedFrameworks() []string {
 	return []string{""}
 }
 
+func (d DotNetDetector) GetApplicationFileInfos(componentPath string, ctx *context.Context) []model.ApplicationFileInfo {
+	// not implemented yet
+	return []model.ApplicationFileInfo{}
+}
+
 // DoFrameworkDetection uses configFilePath to check for the name of the framework
 func (d DotNetDetector) DoFrameworkDetection(language *model.Language, configFilePath string) {
 	framework := getFrameworks(configFilePath)
@@ -52,6 +57,7 @@ func (d DotNetDetector) DoFrameworkDetection(language *model.Language, configFil
 }
 
 func (d DotNetDetector) DoPortsDetection(component *model.Component, ctx *context.Context) {
+	// not implemented yet
 }
 
 func getFrameworks(configFilePath string) string {