--- a/pkg/apis/enricher/framework/java/micronaut_detector.go
+++ b/pkg/apis/enricher/framework/java/micronaut_detector.go
@@ -22,20 +22,25 @@ import (
 
 type MicronautDetector struct{}
 
-type MicronautApplicationProps struct {
-	Micronaut struct {
-		Server struct {
-			Port int `yaml:"port,omitempty"`
-			SSL  struct {
-				Enabled bool `yaml:"enabled,omitempty"`
-				Port    int  `yaml:"port,omitempty"`
-			} `yaml:"ssl,omitempty"`
-		} `yaml:"server,omitempty"`
-	} `yaml:"micronaut,omitempty"`
-}
-
 func (m MicronautDetector) GetSupportedFrameworks() []string {
 	return []string{"Micronaut"}
+}
+
+func (m MicronautDetector) GetApplicationFileInfos(componentPath string, ctx *context.Context) []model.ApplicationFileInfo {
+	return []model.ApplicationFileInfo{
+		{
+			Context: ctx,
+			Root:    componentPath,
+			Dir:     "src/main/resources",
+			File:    "application.yml",
+		},
+		{
+			Context: ctx,
+			Root:    componentPath,
+			Dir:     "src/main/resources",
+			File:    "application.yaml",
+		},
+	}
 }
 
 // DoFrameworkDetection uses the groupId to check for the framework name
@@ -54,26 +59,24 @@ func (m MicronautDetector) DoPortsDetection(component *model.Component, ctx *con
 		return
 	}
 
+	// check if port is set on dockerfile as env var
 	ports = getMicronautPortsFromEnvDockerfile(component.Path)
 	if len(ports) > 0 {
 		component.Ports = ports
 		return
 	}
 
-	bytes, err := utils.ReadAnyApplicationFile(component.Path, []model.ApplicationFileInfo{
-		{
-			Dir:  "src/main/resources",
-			File: "application.yml",
-		},
-		{
-			Dir:  "src/main/resources",
-			File: "application.yaml",
-		},
-	}, ctx)
+	// check source code
+	appFileInfos := m.GetApplicationFileInfos(component.Path, ctx)
+	if len(appFileInfos) == 0 {
+		return
+	}
+	fileBytes, err := utils.GetApplicationFileBytes(appFileInfos[0])
 	if err != nil {
 		return
 	}
-	ports = getMicronautPortsFromBytes(bytes)
+
+	ports = getMicronautPortsFromBytes(fileBytes)
 	if len(ports) > 0 {
 		component.Ports = ports
 	}
@@ -81,7 +84,7 @@ func (m MicronautDetector) DoPortsDetection(component *model.Component, ctx *con
 
 func getMicronautPortsFromBytes(bytes []byte) []int {
 	var ports []int
-	var data MicronautApplicationProps
+	var data model.MicronautApplicationProps
 	err := yaml.Unmarshal(bytes, &data)
 	if err != nil {
 		return []int{}