--- a/pkg/utils/testdata/test.properties
+++ b/pkg/utils/testdata/test.properties
@@ -1,0 +1,3 @@
+key1=value1
+key2=value2
+key3=value3