… docs/public/port_detection.md
… pkg/apis/enricher/framework/java/vertx_detector.go
… pkg/apis/enricher/framework/java/wildfly_detector.go
… pkg/apis/enricher/framework/javascript/nodejs/angular_detector.go
… pkg/apis/enricher/framework/javascript/nodejs/express_detector.go
… pkg/apis/enricher/framework/javascript/nodejs/next_detector.go
… pkg/apis/enricher/framework/javascript/nodejs/nuxt_detector.go
… pkg/apis/enricher/framework/javascript/nodejs/reactjs_detector.go
… pkg/apis/enricher/framework/javascript/nodejs/svelte_detector.go
… pkg/apis/enricher/framework/javascript/nodejs/vue_detector.go
… pkg/apis/enricher/framework/php/laravel_detector.go
… pkg/apis/enricher/framework/python/django_detector.go
… pkg/apis/enricher/framework/python/flask_detector.go
… pkg/apis/model/model.go
… pkg/utils/detector.go
… pkg/utils/detector_test.go