## What does this PR do?

This PR introduces new functionality for every detector in order to ensure a consistent way of picking files for port detection. It also adds more comments, the required unit tests and updates to the documentation.

### Updates
[Consistency]
* The `ApplicationFileInfo` model is expanded to contain the component path and the context. That way it can be the main model used to identify files that may contain information about port detection.

* In every `detector` we add the `GetApplicationFileInfos` which will return specific `ApplicationFileInfo` for every framework.

[Performance]
* For detectors which are looking inside the entire component to find source code files in order detect ports, the `GenerateApplicationFileFromFilters` func has been created. This way we are passing a search keyword in order to go through specific files. For example, we are passing `".go"` for all golang frameworks as, for port detection, we are interested only for golang files.

[Other Updates]
*  As each detector has different functionality for port detection, the `GetApplicationFileContents` and `GetApplicationFileBytes` in order to fetch the content of each application file as `string` or `bytes`.
* All models `types` from detector have been moved to `model.go` and sorted by name. This way we have a specific place to store all model's code and not have them spread around different modules. 
* Further refactoring has been made to `flaskdetector` and `djangodetector` to ensure that we are looping through application files and not picking them one by one.

### Results
1. Performance wise, for a big project with many components (e.g. https://github.com/openshift/hypershift) we can see big difference in the execution time. The main reason is that for many detectors we are filtering the files that we are parsing in order to detect ports. The results:
```bash
# old
$ time ./alizer component ~/github/alizer-tests/hypershift/
...
real	0m4,043s
user	0m4,281s
sys	0m0,368s

# new
time ./alizer component ~/github/alizer-tests/hypershift/
...
real	0m2,556s
user	0m2,909s
sys	0m0,318s
```

2. With the new way of defining potential application files we need only to add an item to every detector's `GetApplicationFileInfos` and not update the whole `DoPortDetection` func.

3. Ensuring that all models are stored in the same place and they have detailed naming. For example, the `ServerXml` from `OpenLibertyDetector` has been moved to `model.go` and renamed to `OpenLibertyServerXml`, in order to be easier identified and found.

### Which issue(s) does this PR fix

fixes https://github.com/devfile/api/issues/1151

### PR acceptance criteria

Testing and documentation do not need to be complete in order for this PR to be approved. We just need to ensure tracking issues are opened.

- [x] Unit/Functional tests

  <!-- _These are run as part of the PR workflow, ensure they are updated_ -->

- [x] Documentation

   <!-- _This includes product docs and READMEs._ -->

### How to test changes / Special notes to the reviewer