@thepetk 2023-09-21
> I was just benchmarking the performance improvements locally (using the same [hypershift](https://github.com/openshift/hypershift) repo) and noticed a strange behavior. When running from the Alizer source code:
> 
> ```
> ☸ ~/w/p/d/alizer on  1151_ft/specific_port_detection [$!?] via 🐹 v1.19.12 
> $ go build -o alizer alizer.go
> 
> ☸ ~/w/p/d/alizer on  1151_ft/specific_port_detection [$!?] via 🐹 v1.19.12 
> $ time ./alizer component /home/asoro/work/tmp/hypershift
> [...]
> ./alizer component /home/asoro/work/tmp/hypershift  18.74s user 1.82s system 123% cpu 16.644 total
> ```
> 
> But when running the same Alizer binary from the [hypershift](https://github.com/openshift/hypershift) repo directory, it takes very long:
> 
> ```
> ☸ ~/w/t/hypershift on  main via 🐹 v1.21.0 
> $ time /home/asoro/work/projects/devfile/alizer/alizer component .
> [...]
> /home/asoro/work/projects/devfile/alizer/alizer component .  142.18s user 3.05s system 111% cpu 2:10.65 total
> ```
> 
> Do you have the same issue? Or am I missing something?

@rm3l no locally I have ~5secs if I run it from the directory of Hypershift and ~3secs if I run it from alizer dir